requirements.txt
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
kiwisolver==1.3.1
matplotlib==3.3.2
networkx==2.5
numpy==1.19.4
Pillow==8.0.1
pycrypto==2.6.1
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
python-dateutil==2.8.1
six==1.15.0
